Fix FPS dropping and optimize buffer streaming
FPS dropping was caused by the list of renderables not being cleared at the
end of each frame. This meant that the list of renderables would accumulate
over every frame, causing objects to be drawn thousands of times.

Buffer streaming was also improved by creating large buffers into which we
append new data every time we want some new buffer space. When the buffer
is filled up, we discard the buffer and create a new one. This prevents the
GPU from stalling as the driver no longer needs to wait for the buffer
memory to be available.